# backend/docs/swagger.yaml
basePath: /
definitions:
...
  logger.LogEntry:
    properties:
      data:
        description: 一行日志原文（含换行符）
        type: string
      index:
        description: 全局递增序号
        type: integer
    type: object
  model.AccessLogItem:
    properties:
      created_at:
        type: string
      headers:
        type: string
      id:
        example: "0"
        type: string
      ip:
        type: string
      latency:
        type: integer
      method:
        type: string
      nickname:
        type: string
      path:
        type: string
      status:
        type: integer
      trace_id:
        type: string
      user_agent:
        type: string
      user_id:
        example: "0"
        type: string
      username:
        type: string
    type: object
  model.AccessLogsResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AccessLogItem'
        type: array
      total:
        type: integer
    type: object
  model.BrowserItem:
    properties:
      browser:
        type: string
      count:
        type: integer
    type: object
...
  model.DBOverviewResponse:
    properties:
      connections:
        type: integer
      name:
        type: string
      size:
        type: string
      table_count:
        type: integer
      type:
        type: string
      version:
        type: string
    type: object
  model.DatabaseInfoResponse:
    properties:
      name:
        type: string
      type:
        type: string
      version:
        type: string
    type: object
...
  model.ExecuteSQLRequest:
    properties:
      sql:
        type: string
    required:
    - sql
    type: object
  model.ExecuteSQLResponse:
    properties:
      affected_rows:
        type: integer
      columns:
        items:
          type: string
        type: array
      execution_time_ms:
        type: integer
      results:
        items:
          additionalProperties: {}
          type: object
        type: array
      type:
        description: '"select" 或 "exec"'
        type: string
    type: object
...
  model.LogsAnalyticsResponse:
    properties:
      browsers:
        items:
          $ref: '#/definitions/model.BrowserItem'
        type: array
      top_users:
        items:
          $ref: '#/definitions/model.TopUserItem'
        type: array
      trend:
        items:
          $ref: '#/definitions/model.TrendItem'
        type: array
    type: object
  model.LogsResponse:
    properties:
      has_more:
        type: boolean
      lines:
        items:
          $ref: '#/definitions/logger.LogEntry'
        type: array
      next_cursor:
        description: 用于加载更早日志的 cursor
        type: integer
    type: object
...
  model.TopUserItem:
    properties:
      count:
        type: integer
      nickname:
        type: string
      user_id:
        example: "0"
        type: string
      username:
        type: string
    type: object
  model.TrendItem:
    properties:
      count:
        type: integer
      date:
        type: string
    type: object
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
info:
  contact:
    name: Wavelet
    url: https://github.com/Rain-kl/Wavelet
  description: Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Wavelet API
  version: 1.0.0
paths:
...
  /api/v1/admin/db-export:
    get:
      description: SQLite 时直接下载 .db 文件；PostgreSQL 时执行 pg_dump 并流式下载 .sql 文件，需要管理员权限
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 数据库文件
          schema:
            type: file
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 导出失败
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 导出数据库
      tags:
      - admin
  /api/v1/admin/db-info:
    get:
      description: 返回当前使用的数据库类型（sqlite/postgres）、名称/路径及版本字符串，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.DatabaseInfoResponse'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取数据库信息
      tags:
      - admin
  /api/v1/admin/db-manage/overview:
    get:
      description: 获取数据库类型、版本、名称、文件大小、表数量及当前连接数，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.DBOverviewResponse'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取数据库运行概览
      tags:
      - admin
  /api/v1/admin/db-manage/query:
    post:
      consumes:
      - application/json
      description: 在当前数据库中执行任意自定义 SQL，如果是查询语句将返回格式化后的列与数据集，否则返回受影响行数，需要管理员权限
      parameters:
      - description: SQL 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ExecuteSQLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 执行完毕
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.ExecuteSQLResponse'
              type: object
        "400":
          description: SQL 语句错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 执行 SQL 查询
      tags:
      - admin
  /api/v1/admin/db-manage/tables:
    get:
      description: 返回当前数据库的所有用户自定义表名称列表，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取数据库所有表名
      tags:
      - admin
  /api/v1/admin/logs:
    get:
      description: 分页获取系统历史日志，cursor=0 获取最新日志，cursor>0 获取更早日志
      parameters:
      - default: 0
        description: 日志游标，0=获取最新
        in: query
        name: cursor
        type: integer
      - default: 200
        description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 日志列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.LogsResponse'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取系统日志
      tags:
      - admin
  /api/v1/admin/logs/access:
    get:
      description: 分页并按照用户、接口路径、时间范围等维度检索用户访问日志列表（需要管理员权限）
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 用户名模糊搜索
        in: query
        name: username
        type: string
      - description: 接口路径模糊搜索
        in: query
        name: path
        type: string
      - description: 起始时间（RFC3339 或 YYYY-MM-DD HH:MM:SS）
        in: query
        name: start_time
        type: string
      - description: 结束时间（RFC3339 或 YYYY-MM-DD HH:MM:SS）
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 访问日志列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.AccessLogsResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Any'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取用户访问日志
      tags:
      - admin
  /api/v1/admin/logs/analytics:
    get:
      description: 聚合统计最近 7 天的每日访问趋势、浏览器分布以及前 10 名最活跃用户排行（需要管理员权限）
      produces:
      - application/json
      responses:
        "200":
          description: 分析统计数据
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/model.LogsAnalyticsResponse'
              type: object
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/response.Any'
        "403":
          description: 无管理员权限
          schema:
            $ref: '#/definitions/response.Any'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取访问日志分析数据
      tags:
      - admin
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"